Fix arrhenius/A in recombination reaction
The rate coefficient for the recombination reaction

Ar+ + e + e -> Ar + e

for use in tps must have units m^6 / mol^2 / s.  The reference used to
set the rates for this case gives it in m^6 / s, so we need to
multiply the pre-exponential coefficient by Avogadro squared (N_A *
N_A).  This was not done originally, resulting in a rate that is too
low by a factor of N_A.  Not surprisingly, that matters.